refactor mockDefaultTabs
