+++ playwright-tests/tests/addons.spec.js
@@ -1,59 +1,6 @@
 const { test, expect } = require("@playwright/test");
 import { pauseIfVideoRecording } from "../testUtils.js";
-import {
-  mockTransactionSubmitRPCResponses,
-  decodeResultJSON,
-  encodeResultJSON,
-} from "../util/transaction.js";
-async function mockTeamsTab(page) {
-  await page.route("https://rpc.mainnet.near.org/", async (route) => {
-    const request = await route.request();
-    const requestPostData = request.postDataJSON();
-
-    if (
-      requestPostData.params &&
-      requestPostData.params.account_id === "devhub.near" &&
-      requestPostData.params.method_name === "get_community"
-    ) {
-      const response = await route.fetch();
-      const json = await response.json();
-
-      const resultObj = decodeResultJSON(json.result.result);
-
-      resultObj.addons = [
-        ...resultObj.addons,
-        {
-          id: "9yhcct",
-          addon_id: "announcements",
-          display_name: "Announcements",
-          enabled: true,
-          parameters: "{}",
-        },
-        {
-          addon_id: "discussions",
-          display_name: "Discussions",
-          enabled: true,
-          id: "gqyrw7",
-          parameters: "{}",
-        },
-        {
-          addon_id: "teams",
-          display_name: "Teams",
-          enabled: true,
-          id: "cqyrw8",
-          parameters: "{}",
-        },
-      ];
-
-      json.result.result = encodeResultJSON(resultObj);
-
-      await route.fulfill({ response, json });
-      return;
-    } else {
-      await route.continue();
-    }
-  });
-}
+import { mockDefaultTabs } from "../util/addons.js";
 
 test.describe("Wallet is not connected", () => {
   test.use({
@@ -76,7 +23,9 @@ test.describe("Wallet is not connected", () => {
   });
 
   test("should load teams addon", async ({ page }) => {
-    await mockTeamsTab(page);
+    await page.route("https://rpc.mainnet.near.org/", async (route) => {
+      await mockDefaultTabs(route);
+    });
 
     await page.goto(
       "/devhub.near/widget/app?page=community&handle=webassemblymusic&tab=teams"
